use size_t in fn sig in lieu of uint16_t

+++ src/extensions/agent_modules/cloud_communication/eg_mqtt_broker_client/src/adu_agent_info_management.c
@@ -76,7 +76,7 @@ void OnMessage_ainfo_resp(
     struct mosquitto* mosq, void* obj, const struct mosquitto_message* msg, const mosquitto_property* props)
 {
     char* correlationData = NULL;
-    uint16_t correlationDataByteLen = 0;
+    size_t correlationDataByteLen = 0;
     ADUC_Retriable_Operation_Context* retriableOperationContext =
         RetriableOperationContextFromAgentInfoMqttLibCallbackUserObj(obj);
     ADUC_AgentInfo_Request_Operation_Data* agentInfoData =
